+++ MOP/src/Vehicles/Managers/Satsuma/SatsumaCustomHoodUse.cs
@@ -18,7 +18,7 @@ using MSCLoader;
 using UnityEngine;
 using MOP.Common;
 
-namespace MOP.Vehicles.Managers
+namespace MOP.Vehicles.Managers.SatsumaManagers
 {
     class SatsumaCustomHoodUse : MonoBehaviour
     {