+++ MOP/src/Vehicles/Managers/Satsuma/SatsumaPartMassManager.cs
@@ -21,7 +21,7 @@ using UnityEngine;
 
 using MOP.Common;
 
-namespace MOP.Vehicles.Managers
+namespace MOP.Vehicles.Managers.SatsumaManagers
 {
     class SatsumaPartMassManager : MonoBehaviour
     {