+++ MOP/src/Vehicles/Managers/Satsuma/SatsumaSeatsManager.cs
@@ -18,7 +18,7 @@ using HutongGames.PlayMaker;
 using UnityEngine;
 using MOP.Common;
 
-namespace MOP.Vehicles.Managers
+namespace MOP.Vehicles.Managers.SatsumaManagers
 {
     class SatsumaSeatsManager : MonoBehaviour
     {